refactor: 에러 처리 파일 구조 및 코드 정리
