+++ src/main/java/com/Alchive/backend/config/ErrorResponse.java
@@ -1,0 +1,32 @@
+package com.Alchive.backend.config;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@Builder
+@RequiredArgsConstructor
+public class ErrorResponse {
+    private final String code;
+    private final String message;
+
+//    @JsonInclude(JsonInclude.Include.NON_EMPTY)
+//    private final List<ValidationErrors> errors;
+//
+//    @Getter
+//    @Builder
+//    @RequiredArgsConstructor
+//    public static class ValidationError {
+//        private final String field;
+//        private final String message;
+//
+//        public static ValidationError of(final FieldError fieldError) {
+//            return ValidationError.builder()
+//                    .field(fieldError.getField())
+//                    .message(fieldError.getDefaultMessage())
+//                    .build();
+//        }
+//    }
+
+}
